SPRT is fully functional
